selection

+++ 2-selection_sort.c
@@ -19,7 +19,7 @@ void swap_e(int *a, int *b)
  */
 void selection_sort(int *array, size_t size)
 {
-	int i, j, min_index;
+	size_t i, j, min_index;
 
 	for (i = 0; i < size - 1; i++)
 	{
